add en_US, en_CA, fr_CA (move upc related items to en_US)
