+++ faker/providers/barcode/fr_CA/__init__.py
@@ -1,0 +1,5 @@
+from ..en_CA import Provider as BarcodeProvider
+
+
+class Provider(BarcodeProvider):
+    pass
